Add focus/blur logic for useValue
When a new value is received for an input that has its value updated through useValue we want to set the internal state value to the external value when the field is blurred. This is what happens if a different user is updating a field the current user is not editing. If we don't do this the value of the other user will be overwritten with the old value.

If the user has the field active, we want to preserve the local value and persist it to firestore – overwriting whatever is there.