Rework branch content creation
Branches are special in the sense that they contain refs to other nodes that make up their content. In order
to create preset branches for negative content and extra informatino we need to create information / error +
result nodes in the same operation as creating the branch.

This change alters a few types and the addNode function in order to allow passing PageContent objects as either
an OrderedMap of node references *or* an array of PageContent objects. In the latter case the nodes and references
are created in the addNodes function.