Fix ACL issue when getting in the getImageUrl function
When running in the emulator we shouldn't call the getSignedUrl function on the file. It triggers ACL checks even if it's not necessary/possible against an emulator, and an error is thrown. To support running in dev with an emulator we check if it's an emulator and call publicUrl instead.